Decode Avro

+++ src/core/ConsumerRecordBuffer.h
@@ -9,14 +9,14 @@ namespace core {
 class ConsumerRecordBuffer
 {
 public:
-    explicit ConsumerRecordBuffer(size_t size);
+    explicit ConsumerRecordBuffer(int64_t size);
 
     void push(std::unique_ptr<ConsumerRecord> &&record);
     ConsumerRecords records();
 
 private:
-    size_t m_size;
-    size_t m_write;
+    int64_t m_size;
+    int64_t m_write;
     std::mutex m_mu;
     boost::circular_buffer<std::unique_ptr<ConsumerRecord>> m_buffer;
 };
